+++ yas/src/core/game/mod.rs
@@ -1,11 +1,11 @@
-use super::{
-    scanner::{ItemScanner, ScannerCore},
-    *,
-};
+use super::*;
 use std::{fmt::*, ops::DerefMut};
 
 pub mod genshin;
 pub mod starrail;
+
+mod scanner;
+pub use scanner::*;
 
 type R = RectBound<WindowInfoType>;
 type P = Pos<WindowInfoType>;
@@ -29,11 +29,6 @@ pub enum Item {
     StarrailRelic(starrail::StarrailRelic),
 }
 
-pub enum Scanner {
-    Genshin(genshin::YasGenshinScanner),
-    StarRail(starrail::YasStarRailScanner),
-}
-
 #[derive(Debug)]
 pub enum Game {
     Genshin,
@@ -49,14 +44,14 @@ pub fn get_window_info(resolution: Resolution) -> WindowInfo {
     }
 }
 
-pub fn parse_level(str_level: &str) -> u32 {
+pub fn parse_level(str_level: &str) -> u8 {
     let pos = str_level.find('+');
 
     if pos.is_none() {
-        return str_level.parse::<u32>().unwrap();
+        return str_level.parse::<u8>().unwrap();
     }
 
-    str_level[0..pos.unwrap()].parse::<u32>().unwrap()
+    str_level[0..pos.unwrap()].parse::<u8>().unwrap()
 }
 
 impl WindowInfo {
@@ -69,31 +64,6 @@ impl WindowInfo {
         match self {
             WindowInfo::Genshin(info) => ScanInfo::Genshin(info.to_scan_info(float_size)),
             WindowInfo::StarRail(info) => ScanInfo::StarRail(info.to_scan_info(float_size)),
-        }
-    }
-}
-
-impl Scanner {
-    pub fn new(
-        scan_info: ScanInfo,
-        config: YasScannerConfig,
-        game_info: GameInfo,
-        model: &[u8],
-        content: &str,
-    ) -> Self {
-        let core = ScannerCore::new(scan_info, config, game_info, model, content);
-
-        match crate::TARGET_GAME.get().unwrap() {
-            Game::Genshin => Scanner::Genshin(genshin::YasGenshinScanner(core)),
-            Game::StarRail => Scanner::StarRail(starrail::YasStarRailScanner(core)),
-            _ => crate::error_and_quit!("不支持的游戏类型"),
-        }
-    }
-
-    pub fn data(&self) -> &ScannerCore {
-        match self {
-            Scanner::Genshin(scanner) => &scanner.0,
-            Scanner::StarRail(scanner) => &scanner.0,
         }
     }
 }
@@ -142,28 +112,6 @@ impl DerefMut for ScanInfo {
         match self {
             ScanInfo::StarRail(info) => &mut info.shared,
             ScanInfo::Genshin(info) => &mut info.shared,
-        }
-    }
-}
-
-impl Deref for Scanner {
-    type Target = dyn ItemScanner;
-
-    #[inline]
-    fn deref(&self) -> &Self::Target {
-        match self {
-            Scanner::Genshin(scanner) => scanner,
-            Scanner::StarRail(scanner) => scanner,
-        }
-    }
-}
-
-impl DerefMut for Scanner {
-    #[inline]
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        match self {
-            Scanner::Genshin(scanner) => scanner,
-            Scanner::StarRail(scanner) => scanner,
         }
     }
 }
